resolved conflicts on fetch aftr creating updateleavetype wrks

+++ src/services/WolfDen.Application/DTOs/LeaveManagement/LeaveBalanceDto.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using WolfDen.Domain.Entity;
 
 namespace WolfDen.Application.DTOs.LeaveManagement
 {
     public class LeaveBalanceDto
     {
         public string Name { get; set; }
-        public int Balance { get; set; }
+        public decimal Balance { get; set; }
     }
 }
