+++ src/services/WolfDen.Domain/Entity/LeaveIncrementLog.cs
@@ -5,12 +5,12 @@
         public int Id {  get;private set; }
         public int LeaveBalanceId { get; private set; }
         public DateOnly LogDate { get; private set; }
-        public int CurrentBalance { get; private set; }
+        public decimal CurrentBalance { get; private set; }
         public int IncrementValue { get; private set; }
         public virtual LeaveBalance LeaveBalance { get; private set; }
         private LeaveIncrementLog() {  }
 
-        public LeaveIncrementLog(int leaveBalanceId, DateOnly logDate, int currentBalance, int incrementValue)
+        public LeaveIncrementLog(int leaveBalanceId, DateOnly logDate, decimal currentBalance, int incrementValue)
         {
             LeaveBalanceId = leaveBalanceId;
             LogDate = logDate;
